requirements.txt
streamlit>=1.32.0
streamlit-extras>=0.4.0
chromadb>=0.4.22
sentence-transformers>=2.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
ollama>=0.1.6
pymupdf>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tiktoken>=0.5.0
